final commit for now, app completed, working on deployement

+++ client/ResumeApp/src/app/dashboard/dashboard.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { ApplicantService } from '../services/applicant.service';
 import { Applicant } from '../applicant.class';
 import { AgGridAngular } from 'ag-grid-angular';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-dashboard',
@@ -15,27 +16,41 @@ export class DashboardComponent implements OnInit {
   applicants: Array<Applicant>
 
   columnDefs = [
-    {  field: 'id', checkboxSelection: true },
-    {  field: 'firstName', sortable: true, filter: true },
-    {  field: 'lastName', sortable: true, filter: true },
-    {  field: 'email', sortable: true, filter: true },
-    {  field: 'position', sortable: true, filter: true }
+    { field: 'id', sortable: true, checkboxSelection: true },
+    { field: 'firstName', sortable: true, filter: true },
+    { field: 'lastName', sortable: true, filter: true },
+    { field: 'email', sortable: true, filter: true },
+    { field: 'position', sortable: true, filter: true },
+    { field: 'resume', cellRenderer(data) {
+        return `<a href=${data.value} target="_blank">Download Resume</a>`
+    } }
   ]
 
-  constructor(private applicantService: ApplicantService) { }
+  constructor(private applicantService: ApplicantService, private snackBar: MatSnackBar) { }
 
   ngOnInit() {
     this.applicantService.fetchApplicants().subscribe(res=> {
       this.applicants = res.applicants
-      console.log(this.applicants)
     })
   }
 
   deleteSelectedApplicants() {
     const selectedData: Array<Applicant> = this.agGrid.api.getSelectedNodes().map(node => node.data)
     const selectedIds = selectedData.map(applicant=> parseInt(applicant.id))
+    if( !selectedIds.length ) {
+      this.snackBar.open('No applicant selected!', 'Dismiss', {
+        duration: 2000
+      })
+      return
+    }
     this.applicantService.deleteApplicant(selectedIds).subscribe(res=> {
-      console.log('Deleted!')
+      this.snackBar.open(`Deleted ${res} applicants`, 'Dismiss', {
+        duration: 2000
+      })
+      this.applicantService.fetchApplicants().subscribe(res=> {
+        this.applicants = res.applicants
+        console.log(this.applicants)
+      })
     })
   }
